Enable watchdog
At maximum scaling, the watchdog will expire after 131 seconds.  The
sleep instruction resets this counter, so CAN bus activity implicitly
resets the timer.  Enabling the watchdog means if we go 2 minutes with
zero CAN activity, we'll reset, which will turn off any outputs.

This is intended as a safeguard for when the key-off CAN bus message is
missed.